+++ src/Resources/ResourceManager/Implementation/Policy/NewAzurePolicyAssignment.cs
@@ -121,8 +121,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         [ValidateNotNullOrEmpty]
         public PolicyAssignmentEnforcementMode? EnforcementMode { get; set; }
 
-        [CmdletParameterBreakingChange("AssignIdentity", ChangeDescription = "Parameter AssignIdentity is invalid and preserved for compatibility.")]
-        [ValidateNotNullOrEmpty]
+        [CmdletParameterBreakingChange("AssignIdentity", ChangeDescription = "Parameter AssignIdentity is deprecated and will be removed in future releases. Please use the 'IdentityType' parameter instead.")]
+        [Parameter(Mandatory = false, HelpMessage = PolicyHelpStrings.PolicyAssignmentAssignIdentityHelp)]
         public SwitchParameter AssignIdentity { get; set; }
 
         /// <summary>
@@ -213,7 +213,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         /// </summary>
         private bool CheckIfIdentityPresent()
         {
-            if (this.IdentityType != null & this.Location == null)
+            if (this.IdentityType != null & this.Location == null || this.AssignIdentity != null && this.AssignIdentity.IsPresent && this.Location == null)
             {
                 throw new PSInvalidOperationException("Location needs to be specified if a managed identity is to be assigned to the policy assignment.");
             }
@@ -236,6 +236,11 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         /// </summary>
         private JToken GetResource()
         {
+            if (this.AssignIdentity != null && this.AssignIdentity.IsPresent)
+            {
+                this.IdentityType = ManagedIdentityType.SystemAssigned;
+            }
+
             ResourceIdentity identityObject = this.IdentityType != null ?
                 (this.IdentityType == ManagedIdentityType.SystemAssigned ?
                 new ResourceIdentity { Type = ManagedIdentityType.SystemAssigned.ToString() } :