+++ src/Resources/Resources/ChangeLog.md
@@ -23,7 +23,7 @@
 * [Breaking change] Changed the returned `Id` in PSDenyAssignment from GUID string to fully qualified ID
 * Allowed parameter `Id` in `Get-AzDenyAssignment` to accept fully qualified ID
 * Added new cmdlet `Publish-AzBicepModule` for publishing Bicep modules
-* Removed support for `AssignIdentity` parameter in `*-AzPolicyAssignment` cmdlets.
+* Added deprecation message for `AssignIdentity` parameter in `*-AzPolicyAssignment` cmdlets.
 * Added support for user assigned managed identities in policy assignments by adding `IdentityType` and `IdentityId` parameters to `*-AzPolicyAssignment` cmdlets.
 * Updated policy cmdlets to use new api version 2021-06-01 that introduces support for user assigned managed identities in policy assignments.